requirements.txt

# Usage:
#conda create -n anaconda_env
#conda activate anaconda_env
#python3 -m venv anaconda_env
#pip install -r requirements.txt

# Base ------------------------------------------------------------------------
gitpython>=3.1.31
matplotlib>=3.7.1
numpy>=1.24.3
opencv-python>=4.7.0.72
Pillow>=9.5.0
psutil  # system resources
PyYAML>=5.3.1
requests>=2.28.2
scipy>=1.10.1
thop>=0.1.1  # FLOPs computation
torch>=2.0.0  # see https://pytorch.org/get-started/locally (recommended)
torchvision>=0.8.1
tqdm>=4.65.0
# protobuf<=3.20.1  # https://github.com/ultralytics/yolov5/issues/8012

# Logging ---------------------------------------------------------------------
# tensorboard>=2.4.1
# clearml>=1.2.0
# comet

# Plotting --------------------------------------------------------------------
pandas>=2.0.0
seaborn>=0.12.2

# Export ----------------------------------------------------------------------
# coremltools>=6.0  # CoreML export
# onnx>=1.10.0  # ONNX export
# onnx-simplifier>=0.4.1  # ONNX simplifier
# nvidia-pyindex  # TensorRT export
# nvidia-tensorrt  # TensorRT export
# scikit-learn<=1.1.2  # CoreML quantization
# tensorflow>=2.4.0  # TF exports (-cpu, -aarch64, -macos)
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev  # OpenVINO export

# Deploy ----------------------------------------------------------------------
setuptools>=65.5.1 # Snyk vulnerability fix
# tritonclient[all]~=2.24.0

# Extras ----------------------------------------------------------------------
# ipython  # interactive notebook
# mss  # screenshots
# albumentations>=1.0.3
# pycocotools>=2.0.6  # COCO mAP
# ultralytics  # HUB https://hub.ultralytics.com